refactor(VAutocomplete): implement using VCombobox
